refactor(select): Barrel file renaming
